Move sinks/sources to path structure that matches their packages
