+++ instrumentation/orchestrion.tool.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	_ "github.com/AikidoSec/firewall-go/instrumentation/sinks/databasesql"
+	_ "github.com/AikidoSec/firewall-go/instrumentation/sinks/database/sql"
 	_ "github.com/AikidoSec/firewall-go/instrumentation/sinks/os"
-	_ "github.com/AikidoSec/firewall-go/instrumentation/sources/gingonic"
-	_ "github.com/AikidoSec/firewall-go/instrumentation/sources/labstackecho"
+	_ "github.com/AikidoSec/firewall-go/instrumentation/sources/gin-gonic/gin"
+	_ "github.com/AikidoSec/firewall-go/instrumentation/sources/labstack/echo"
 )
