+++ instrumentation/sinks/database/sql/main.go
@@ -1,4 +1,4 @@
-package databasesql
+package sql
 
 import (
 	"context"
